// src/app/main-modules/basicInformation/lubricant/model/lubricantDto.ts
export namespace LubricantDto {

    export class Create {
        title: string;
        type: string;

    }

    export class Update {
        id: string;
        title: string;
        type: string;
    }

    export class GetPageSearchDto {
        id: string;
        title: string;
        type: string;
    }

    export class GetPage {

        id: string;
        title: string;
        type: string;


    }

    export class GetOne {

        id: string;
        title: string;
        type: string;

    }

    export class GetAll {
        id: string;
        title: string;
        type: string;

    }


}